fixed validator issue

+++ kcommons/src/main/java/com/kairos/dto/activity/activity/activity_tabs/communication_tab/FrequencySettings.java
@@ -8,18 +8,18 @@ import java.io.Serializable;
  * CreatedBy vipulpandey on 6/10/18
  **/
 public class FrequencySettings implements Serializable {
-    private Short timeValue;
+    private Integer timeValue;
     private DurationType durationType;
 
     public FrequencySettings() {
 
     }
 
-    public Short getTimeValue() {
+    public Integer getTimeValue() {
         return timeValue;
     }
 
-    public void setTimeValue(Short timeValue) {
+    public void setTimeValue(Integer timeValue) {
         this.timeValue = timeValue;
     }
